+++ changelog.md
@@ -2,11 +2,29 @@
 
 ## Version 5
 
+### Version 5.0.1
+
 * `quasar-utils`: Upgrade quasar to version 0.8.0 (JDK 11 support)
+
+* `api-scanner`: Update to support Gradle's `java-library` plugin.
+
+* `quasar-utils`: Add a `quasar` extension so that we can exclude packages from being instrumented by the Quasar agent. Also expose Quasar's `verbose` and `debug` options using extension properties. And `group`, `version` and `classifier` properties so that we can configure whicn agent artifact to use.
+
+* `quasar-utils`: Remove Quasar's transitive dependencies from Gradle's runtime classpath.
 
 ### Version 5.0.0
 
+* Upgrade to Gradle 5.4.1 / Kotlin 1.3.21.
+
+* `quasar-utils`: Apply the `java` plugin automatically.
+
+## Version 4
+
+### Version 4.0.44
+
 * `cordapp`: Explicitly configure all of the CorDapp's `jar` task properties to ensure the CorDapp has a stable SHA256 hash.
+
+### Version 4.0.43
 
 * `cordformation`: Add support for network parameter overrides.
 
@@ -16,15 +34,15 @@
 
 * `api-scanner`: Compatibility fix for Gradle 5.x.
 
-* `cordformation`: Unload `SecurityProvider` instances installed by Network Bootstrapper to prevent linkage exceptions later.
-
-* `cordformation`: Revert `deployNodes` back to using the runtime classpath again. Anything needed solely by `deployNodes` can be added to Gradle's `cordaRuntime` configuration instead.
+### Version 4.0.42
 
 * `cordformation`: Add Jolokia agent to the `cordaRuntime` configuration to prevent the `cordapp` plugin from including it inside the CorDapp.
 
-* `quasar-utils`: Apply the `java` plugin automatically.
+### Version 4.0.41
 
-## Version 4
+* `cordformation`: Unload `SecurityProvider` instances installed by Network Bootstrapper to prevent linkage exceptions later.
+
+* `cordformation`: Revert `deployNodes` back to using the runtime classpath again. Anything needed solely by `deployNodes` can be added to Gradle's `cordaRuntime` configuration instead.
 
 ### Version 4.0.40
 
@@ -94,7 +112,7 @@
 
 ### Version 4.0.25
 
-* `quasar-utils': The plugin now does not require projects to specify `quasar_group` and `qauas
+* `quasar-utils`: The plugin now does not require projects to specify `quasar_group` and `quasar_version`
 
 ### Version 4.0.24